# EventEdge/openapi.yaml
openapi: 3.0.3
info:
  title: EventEdge API
  version: 0.1.0
  description: |
    API-first contract for the EventEdge MVP.

    All timestamps are UTC RFC 3339. Published signals are immutable: a recalculation
    creates a new revision and marks the previous revision as superseded.
  contact:
    name: EventEdge team
servers:
  - url: https://api.eventedge.example
    description: Production placeholder
...
security:
  - bearerAuth: []
paths:
  /v1/signals:
    get:
      tags: [Signals]
      summary: List signal revisions
      operationId: listSignals
      parameters:
        - $ref: '#/components/parameters/TickerQuery'
        - name: direction
          in: query
          description: Filter by one or more directions.
          schema:
            type: array
            items:
              $ref: '#/components/schemas/SignalDirection'
          style: form
          explode: false
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/SignalStatus'
        - name: min_confidence
          in: query
          schema:
            type: number
            format: double
            minimum: 0
            maximum: 1
        - name: active_at
          in: query
          description: Return signals active at this instant. Defaults to now.
          schema:
            type: string
            format: date-time
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/IfNoneMatch'
      responses:
        '200':
          description: A page ordered by `as_of DESC, id DESC`.
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SignalListResponse'
        '304':
          description: The cached representation is still current.
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
      x-yc-apigateway-integration:
        $ref: '#/components/x-yc-apigateway-integrations/api'
  /v1/signals/{signal_id}:
    parameters:
      - $ref: '#/components/parameters/SignalId'
    get:
      tags: [Signals]
      summary: Get one immutable signal revision
      operationId: getSignal
      parameters:
        - $ref: '#/components/parameters/IfNoneMatch'
      responses:
        '200':
          description: Signal revision.
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SignalResponse'
        '304':
          description: The cached representation is still current.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
      x-yc-apigateway-integration:
        $ref: '#/components/x-yc-apigateway-integrations/api'
...
  /v1/events:
    get:
      tags: [Events]
      summary: List normalized news events
      operationId: listEvents
      parameters:
        - $ref: '#/components/parameters/TickerQuery'
        - name: published_from
          in: query
          schema:
            type: string
            format: date-time
        - name: published_to
          in: query
          schema:
            type: string
            format: date-time
        - name: min_materiality
          in: query
          schema:
            type: number
            format: double
            minimum: 0
            maximum: 1
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: A page ordered by `event_time DESC, id DESC`.
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
      x-yc-apigateway-integration:
        $ref: '#/components/x-yc-apigateway-integrations/api'
  /v1/events/{event_id}:
    parameters:
      - $ref: '#/components/parameters/EventId'
    get:
      tags: [Events]
      summary: Get a normalized event and its evidence
      operationId: getEvent
      responses:
        '200':
          description: Normalized event.
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
      x-yc-apigateway-integration:
        $ref: '#/components/x-yc-apigateway-integrations/api'
...
  /v1/internal/news:
    post:
      tags: [Internal ingestion]
      summary: Submit one source document for asynchronous processing
      operationId: ingestNews
      x-required-role: ingestor
      security:
        - serviceApiKey: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewsIngestRequest'
      responses:
        '202':
          description: The document was accepted or previously accepted with the same idempotency key.
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
            Location:
              description: URL of the asynchronous job.
              schema:
                type: string
                format: uri-reference
            Idempotency-Replayed:
              description: Present and equal to true when the same command was accepted earlier.
              schema:
                type: boolean
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '429':
          $ref: '#/components/responses/RateLimited'
      x-yc-apigateway-integration:
        $ref: '#/components/x-yc-apigateway-integrations/api'
  /v1/jobs/{job_id}:
    parameters:
      - $ref: '#/components/parameters/JobId'
    get:
      tags: [Jobs]
      summary: Get asynchronous operation status
      operationId: getJob
      responses:
        '200':
          description: Current job state.
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
            Retry-After:
              description: Suggested polling delay in seconds while the job is not terminal.
              schema:
                type: integer
                minimum: 1
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
      x-yc-apigateway-integration:
        $ref: '#/components/x-yc-apigateway-integrations/api'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    serviceApiKey:
      type: apiKey
      in: header
      name: X-API-Key
...
  headers:
    XRequestId:
      description: End-to-end request identifier.
      schema:
        type: string
        minLength: 8
        maxLength: 128
    ETag:
      description: Version of the returned representation.
      schema:
        type: string
  parameters:
    SignalId:
      name: signal_id
      in: path
      required: true
      schema:
        type: string
        pattern: '^sig_[0-9A-HJKMNP-TV-Z]{26}$'
    EventId:
      name: event_id
      in: path
      required: true
      schema:
        type: string
        pattern: '^evt_[0-9A-HJKMNP-TV-Z]{26}$'
    JobId:
      name: job_id
      in: path
      required: true
      schema:
        type: string
        pattern: '^job_[0-9A-HJKMNP-TV-Z]{26}$'
...
    TickerQuery:
      name: ticker
      in: query
      schema:
        type: string
        pattern: '^[A-Z0-9]{1,12}$'
        example: SBER
    Cursor:
      name: cursor
      in: query
      description: Opaque cursor returned in `meta.next_cursor`.
      schema:
        type: string
        maxLength: 1024
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: Stable client-generated key retained for at least 24 hours.
      schema:
        type: string
        minLength: 8
        maxLength: 128
    IfNoneMatch:
      name: If-None-Match
      in: header
      required: false
      schema:
        type: string
...
  responses:
    BadRequest:
      description: Request syntax or parameters are invalid.
      headers:
        X-Request-Id:
          $ref: '#/components/headers/XRequestId'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Unauthorized:
      description: Authentication is missing or invalid.
      headers:
        X-Request-Id:
          $ref: '#/components/headers/XRequestId'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
    NotFound:
      description: Resource does not exist or is not visible to the principal.
      headers:
        X-Request-Id:
          $ref: '#/components/headers/XRequestId'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Conflict:
      description: The command conflicts with existing resource state or idempotency data.
      headers:
        X-Request-Id:
          $ref: '#/components/headers/XRequestId'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
    PayloadTooLarge:
      description: Request body exceeds the configured ingestion limit.
      headers:
        X-Request-Id:
          $ref: '#/components/headers/XRequestId'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
    RateLimited:
      description: Request rate or monthly processing allowance was exceeded.
      headers:
        X-Request-Id:
          $ref: '#/components/headers/XRequestId'
        Retry-After:
          description: Suggested retry delay in seconds.
          schema:
            type: integer
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
  schemas:
    SignalDirection:
      type: string
      enum: [up, neutral, down]
    SignalAction:
      type: string
      enum: [consider_buy, no_action, review_position]
    SignalStatus:
      type: string
      enum: [active, expired, superseded, invalidated]
    Horizon:
      type: object
      additionalProperties: false
      required: [value, unit]
      properties:
        value:
          type: integer
          minimum: 1
          maximum: 90
        unit:
          type: string
          enum: [hours, calendar_days, trading_days]
    FactorContribution:
      type: object
      additionalProperties: false
      required: [code, label, contribution]
      properties:
        code:
          type: string
          pattern: '^[a-z][a-z0-9_]{1,63}$'
        label:
          type: string
          maxLength: 120
        contribution:
          type: number
          format: double
          minimum: -1
          maximum: 1
    Signal:
      type: object
      additionalProperties: false
      required:
        - id
        - ticker
        - as_of
        - data_cutoff_at
        - status
        - direction
        - action
        - horizon
        - score
        - strength
        - confidence
        - summary
        - factor_contributions
        - evidence_refs
        - expires_at
        - invalidation_conditions
        - model_version
        - config_version
        - created_at
      properties:
        id:
          type: string
          example: sig_01JZK6K5GDX90Q2X8C0R4D7M9P
        previous_revision_id:
          type: string
          nullable: true
        ticker:
          type: string
          example: SBER
        as_of:
          type: string
          format: date-time
        data_cutoff_at:
          type: string
          format: date-time
        status:
          $ref: '#/components/schemas/SignalStatus'
        direction:
          $ref: '#/components/schemas/SignalDirection'
        action:
          $ref: '#/components/schemas/SignalAction'
        horizon:
          $ref: '#/components/schemas/Horizon'
        score:
          type: number
          format: double
          minimum: -100
          maximum: 100
        strength:
          type: number
          format: double
          minimum: 0
          maximum: 1
        confidence:
          type: number
          format: double
          minimum: 0
          maximum: 1
        summary:
          type: string
          minLength: 1
          maxLength: 800
        factor_contributions:
          type: array
          maxItems: 20
          items:
            $ref: '#/components/schemas/FactorContribution'
        evidence_refs:
          type: array
          maxItems: 50
          items:
            type: string
        expires_at:
          type: string
          format: date-time
        invalidation_conditions:
          type: array
          maxItems: 10
          items:
            type: string
            maxLength: 300
        model_version:
          type: string
          example: baseline-0.1.0
        config_version:
          type: integer
          minimum: 1
        created_at:
          type: string
          format: date-time
    Evidence:
      type: object
      additionalProperties: false
      required: [news_id, source_id, published_at, title, url]
      properties:
        news_id:
          type: string
        source_id:
          type: string
        published_at:
          type: string
          format: date-time
        title:
          type: string
          maxLength: 500
        url:
          type: string
          format: uri
        content_hash:
          type: string
          description: Hash used to prove which source representation was analyzed.
    Event:
      type: object
      additionalProperties: false
      required:
        - id
        - event_time
        - detected_at
        - event_type
        - title
        - summary
        - tickers
        - novelty
        - materiality
        - evidence
        - extractor_version
      properties:
        id:
          type: string
          example: evt_01JZK5Y4V2JABR9N4F0S7Q2XWC
        event_time:
          type: string
          format: date-time
        detected_at:
          type: string
          format: date-time
        event_type:
          type: string
          maxLength: 80
        title:
          type: string
          maxLength: 500
        summary:
          type: string
          maxLength: 1500
        tickers:
          type: array
          maxItems: 50
          items:
            type: string
        novelty:
          type: number
          format: double
          minimum: 0
          maximum: 1
        materiality:
          type: number
          format: double
          minimum: 0
          maximum: 1
        evidence:
          type: array
          minItems: 1
          maxItems: 50
          items:
            $ref: '#/components/schemas/Evidence'
        extractor_version:
          type: string
        quality_status:
          type: string
          enum: [complete, partial, conflicting]
          default: complete
...
    JobStatus:
      type: string
      enum: [queued, running, succeeded, failed, cancelled]
    Job:
      type: object
      additionalProperties: false
      required: [id, kind, status, created_at, updated_at]
      properties:
        id:
          type: string
          example: job_01JZK7AHMXYCG6A5T9D1B8R3QP
        kind:
          type: string
          enum: [news_ingestion, extraction, signal_recalculation, backfill, source_validation]
        status:
          $ref: '#/components/schemas/JobStatus'
        progress:
          type: number
          format: double
          minimum: 0
          maximum: 1
          nullable: true
        result_ref:
          type: string
          nullable: true
        error:
          allOf:
            - $ref: '#/components/schemas/Problem'
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
          nullable: true
    NewsIngestRequest:
      type: object
      additionalProperties: false
      required: [source_id, external_id, published_at, title, url, content]
      properties:
        source_id:
          type: string
          pattern: '^[a-z][a-z0-9_-]{2,63}$'
        external_id:
          type: string
          minLength: 1
          maxLength: 500
        published_at:
          type: string
          format: date-time
        received_at:
          type: string
          format: date-time
        title:
          type: string
          minLength: 1
          maxLength: 500
        url:
          type: string
          format: uri
        content:
          type: string
          minLength: 1
          maxLength: 200000
        language:
          type: string
          default: ru
          enum: [ru, en]
        source_metadata:
          type: object
          additionalProperties: true
          maxProperties: 50
...
    PageMeta:
      type: object
      additionalProperties: false
      required: [limit, has_more]
      properties:
        limit:
          type: integer
        has_more:
          type: boolean
        next_cursor:
          type: string
          nullable: true
    SignalResponse:
      type: object
      additionalProperties: false
      required: [data]
      properties:
        data:
          $ref: '#/components/schemas/Signal'
    SignalListResponse:
      type: object
      additionalProperties: false
      required: [data, meta]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Signal'
        meta:
          $ref: '#/components/schemas/PageMeta'
    EventResponse:
      type: object
      additionalProperties: false
      required: [data]
      properties:
        data:
          $ref: '#/components/schemas/Event'
    EventListResponse:
      type: object
      additionalProperties: false
      required: [data, meta]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Event'
        meta:
          $ref: '#/components/schemas/PageMeta'
...
    JobResponse:
      type: object
      additionalProperties: false
      required: [data]
      properties:
        data:
          $ref: '#/components/schemas/Job'
...
    Problem:
      type: object
      additionalProperties: false
      required: [type, title, status, code, request_id]
      properties:
        type:
          type: string
          format: uri-reference
          example: https://api.eventedge.example/problems/invalid-parameter
        title:
          type: string
          example: Invalid request parameter
        status:
          type: integer
          minimum: 400
          maximum: 599
        code:
          type: string
          pattern: '^[A-Z][A-Z0-9_]{2,63}$'
          example: INVALID_PARAMETER
        detail:
          type: string
          maxLength: 2000
        instance:
          type: string
          format: uri-reference
        request_id:
          type: string
        errors:
          type: array
          maxItems: 50
          items:
            type: object
            additionalProperties: false
            required: [location, message]
            properties:
              location:
                type: string
                example: query.min_confidence
              message:
                type: string